Kill in_localaddr(), one less usage of the global list of IPv4 addresses.
This function is used only once in our tree to optimize the size of the
MSS if the forward address correspond to a host on one of our subnets,
but only if ip.mutdisc is disable, which is not the default!

While here get rid of the "#ifdef RTV_MTU", it is here.

ok henning@, mikeb@, bluhm@