When a route to prefix (connected route) is added, if its associated broadcast address is non null, add a broadcast entry flagged with RTF_BROADCAST.
Re-use the existing logic to switch a route to prefix from an ifa
to another to also move this broadcast entry.

Prior to this change broadcast entries were simple clonned ARP entries,
that would be deleted once their timer expired since they would always
be incomplete.

With this change they are now persistant and identifiable with a new flag

This version of the diff prevent a corruption reported by millert@

ok mikeb@, florian@