Fix constraint.
We need a register here, and not a generic memory reference.  We get lucky
with GCC that it chooses a register for "p", but we can't blame clang for
choosing a memory reference if we allow it to do so.

Pointed out by LLVM.

time.c:80:17: error: unexpected token in memory operand

ok kettenis@