Do no change the gateway of local routes for p2p interfaces.
This change was defeating the code in rtrequest1(9) checking for route
entries with the same dst/gw when the same IP address was configured on
multiple interfaces.

As a result, multiple local routes were created for the same address and
marked as multipath.  But changing their gateway to 127.0.0.1 would make
them similar and impossible to remove.  This would leaves entries with a
stall ifa pointer as soon as the address was removed.

Prevent a panic reported by todd@