Drop promiscuously received packets if the trunk(4) interface is not in promiscuous mode.
The long story is that claudio@ had his ssh session reset multiple
times in the hackroom because czarkoff@'s machine was sending reset.
We figured out that the packet was reaching pf because of this missing
check.  pf would then not find any state and sent a reset.

Analyzed with and ok phessler@, claudio@